ProjectBackEnd/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.8.1
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
distro==1.9.0
Django==5.0.3
django-booking==0.7.2
django-cors-headers==4.3.1
django-countries==7.6.1
django-hvad==1.8.0
django-libs==2.0.3
djangorestframework==3.15.1
exceptiongroup==1.2.0
Flask==3.0.3
Flask-Cors==4.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
oauthlib==3.2.2
openai==1.17.0
psycopg==3.1.18
psycopg-binary==3.1.18
pydantic==2.7.0
pydantic_core==2.18.1
pygame==2.5.2
python-dotenv==1.0.1
requests==2.31.0
requests-oauthlib==2.0.0
sniffio==1.3.1
sqlparse==0.4.4
tqdm==4.66.2
typing_extensions==4.10.0
urllib3==2.2.1
Werkzeug==3.0.2
